moving to Decimal adding more info to Symbol

+++ src/account.rs
@@ -1,3 +1,4 @@
+use rust_decimal::Decimal;
 use crate::client::*;
 use crate::errors::*;
 use crate::rest_model::*;
@@ -30,15 +31,15 @@ pub struct OrderRequest {
     #[serde(rename = "type")]
     pub order_type: OrderType,
     pub time_in_force: Option<TimeInForce>,
-    pub quantity: Option<f64>,
-    pub quote_order_qty: Option<f64>,
-    pub price: Option<f64>,
+    pub quantity: Option<Decimal>,
+    pub quote_order_qty: Option<Decimal>,
+    pub price: Option<Decimal>,
     /// A unique id for the order, automatically generated if not sent.
     pub new_client_order_id: Option<String>,
     /// Used with stop loss, stop loss limit, take profit and take profit limit order types.
-    pub stop_price: Option<f64>,
+    pub stop_price: Option<Decimal>,
     /// Used with limit, stop loss limit and take profit limit to create an iceberg order.
-    pub iceberg_qty: Option<f64>,
+    pub iceberg_qty: Option<Decimal>,
     /// Set the response json, market and limit default to full others to ack.
     pub new_order_resp_type: Option<OrderResponse>,
     /// Cannot be greater than 60000
@@ -83,15 +84,15 @@ pub struct CancelReplaceRequest {
     pub order_type: OrderType,
     pub cancel_replace_mode: CancelReplaceMode,
     pub time_in_force: Option<TimeInForce>,
-    pub quantity: Option<f64>,
-    pub quote_order_qty: Option<f64>,
-    pub price: Option<f64>,
+    pub quantity: Option<Decimal>,
+    pub quote_order_qty: Option<Decimal>,
+    pub price: Option<Decimal>,
     pub cancel_new_client_order_id: Option<String>,
     pub cancel_orig_client_order_id: Option<String>,
     pub cancel_order_id: Option<u64>,
     pub new_client_order_id: Option<String>,
-    pub stop_price: Option<f64>,
-    pub iceberg_qty: Option<f64>,
+    pub stop_price: Option<Decimal>,
+    pub iceberg_qty: Option<Decimal>,
     pub new_order_resp_type: Option<OrderResponse>,
     /// Cannot be greater than 60000
     pub recv_window: Option<u64>,
@@ -292,12 +293,13 @@ impl Account {
     /// This methods validates the order request before sending, making sure it complies with Binance rules
     /// # Examples
     /// ```rust,no_run
+    /// use rust_decimal_macros::dec;
     /// use binance::{api::*, account::*, config::*, rest_model::*};
     /// let account: Account = Binance::new_with_env(&Config::testnet());
     /// let limit_buy = OrderRequest {
     ///         symbol: "BTCUSDT".to_string(),
-    ///         quantity: Some(10.0),
-    ///         price: Some(0.014000),
+    ///         quantity: Some(dec!(10.0)),
+    ///         price: Some(dec!(0.014000)),
     ///         order_type: OrderType::Limit,
     ///         side: OrderSide::Buy,
     ///         time_in_force: Some(TimeInForce::FOK),
@@ -319,12 +321,13 @@ impl Account {
     /// This order is sandboxed: it is validated, but not sent to the matching engine.
     /// # Examples
     /// ```rust,no_run
+    /// use rust_decimal_macros::dec;
     /// use binance::{api::*, account::*, config::*, rest_model::*};
     /// let account: Account = Binance::new_with_env(&Config::testnet());
     /// let limit_buy = OrderRequest {
     ///         symbol: "BTCUSDT".to_string(),
-    ///         quantity: Some(10.0),
-    ///         price: Some(0.014000),
+    ///         quantity: Some(dec!(10.0)),
+    ///         price: Some(dec!(0.014000)),
     ///         order_type: OrderType::Limit,
     ///         side: OrderSide::Buy,
     ///         time_in_force: Some(TimeInForce::FOK),
